+++ integration/e2e/e2e_test.go
@@ -132,10 +132,6 @@ func TestAllInOne(t *testing.T) {
 			require.NoError(t, err)
 
 			util.SearchStreamAndAssertTrace(t, grpcClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
-
-			// test websockets
-			wsClient := httpclient.New("ws://"+tempo.Endpoint(3200), "")
-			util.SearchWSStreamAndAssertTrace(t, wsClient, info, now.Add(-20*time.Minute).Unix(), now.Unix())
 		})
 	}
 }
@@ -471,19 +467,6 @@ func callFlush(t *testing.T, ingester *e2e.HTTPService) {
 	res, err := e2e.DoGet("http://" + ingester.Endpoint(3200) + "/flush")
 	require.NoError(t, err)
 	require.Equal(t, http.StatusNoContent, res.StatusCode)
-}
-
-// writeMetrics calls /metrics and write it to text file, useful for debugging e2e tests
-func writeMetrics(t *testing.T, tempo *e2e.HTTPService, filename string) {
-	fmt.Printf("Calling /metrics on %s\n", tempo.Name())
-	res, err := e2e.DoGet("http://" + tempo.Endpoint(3200) + "/metrics")
-	require.NoError(t, err)
-	require.Equal(t, http.StatusOK, res.StatusCode)
-
-	body, err := io.ReadAll(res.Body)
-	require.NoError(t, err)
-	err = os.WriteFile("metrics_"+filename+"_dump.txt", body, 0o644)
-	require.NoError(t, err)
 }
 
 func callIngesterRing(t *testing.T, svc *e2e.HTTPService) {