+++ modules/frontend/frontend.go
@@ -14,10 +14,10 @@ import (
 	"github.com/golang/protobuf/jsonpb" //nolint:all //deprecated
 	"github.com/golang/protobuf/proto"  //nolint:all //deprecated
 	"github.com/grafana/dskit/user"
-	"github.com/grafana/tempo/modules/frontend/combiner"
 	"github.com/opentracing/opentracing-go"
 	"github.com/prometheus/client_golang/prometheus"
 
+	"github.com/grafana/tempo/modules/frontend/combiner"
 	"github.com/grafana/tempo/modules/overrides"
 	"github.com/grafana/tempo/pkg/api"
 	"github.com/grafana/tempo/pkg/cache"
@@ -28,7 +28,7 @@ import (
 type streamingSearchHandler func(req *tempopb.SearchRequest, srv tempopb.StreamingQuerier_SearchServer) error
 
 type QueryFrontend struct {
-	TraceByIDHandler, SearchHandler, SpanMetricsSummaryHandler, SearchWSHandler                http.Handler
+	TraceByIDHandler, SearchHandler, SpanMetricsSummaryHandler, QueryRangeHandler              http.Handler
 	SearchTagsHandler, SearchTagsV2Handler, SearchTagsValuesHandler, SearchTagsValuesV2Handler http.Handler
 	cacheProvider                                                                              cache.Provider
 	streamingSearch                                                                            streamingSearchHandler
@@ -84,10 +84,13 @@ func New(cfg Config, next http.RoundTripper, o overrides.Interface, reader tempo
 	searchTagsValuesV2Middleware := MergeMiddlewares(
 		newMultiTenantMiddleware(cfg, combiner.NewSearchTagValuesV2, logger),
 		newSearchTagsMiddleware(), retryWare)
-
-	spanMetricsMiddleware := MergeMiddlewares(
+	metricsMiddleware := MergeMiddlewares(
 		newMultiTenantUnsupportedMiddleware(cfg, logger),
-		newSpanMetricsMiddleware(), retryWare)
+		newMetricsMiddleware(), retryWare)
+
+	queryRangeMiddleware := MergeMiddlewares(
+		newMultiTenantUnsupportedMiddleware(cfg, logger),
+		newQueryRangeMiddleware(cfg, o, reader, logger), retryWare)
 
 	traces := traceByIDMiddleware.Wrap(next)
 	search := searchMiddleware.Wrap(next)
@@ -96,7 +99,8 @@ func New(cfg Config, next http.RoundTripper, o overrides.Interface, reader tempo
 	searchTagValues := searchTagsValuesMiddleware.Wrap(next)
 	searchTagValuesV2 := searchTagsValuesV2Middleware.Wrap(next)
 
-	metrics := spanMetricsMiddleware.Wrap(next)
+	metrics := metricsMiddleware.Wrap(next)
+	queryrange := queryRangeMiddleware.Wrap(next)
 
 	return &QueryFrontend{
 		TraceByIDHandler:          newHandler(traces, traceByIDSLOPostHook(cfg.TraceByID.SLO), nil, logger),
@@ -105,9 +109,8 @@ func New(cfg Config, next http.RoundTripper, o overrides.Interface, reader tempo
 		SearchTagsV2Handler:       newHandler(searchTagsV2, nil, nil, logger),
 		SearchTagsValuesHandler:   newHandler(searchTagValues, nil, nil, logger),
 		SearchTagsValuesV2Handler: newHandler(searchTagValuesV2, nil, nil, logger),
-
 		SpanMetricsSummaryHandler: newHandler(metrics, nil, nil, logger),
-		SearchWSHandler:           newSearchStreamingWSHandler(cfg, o, retryWare.Wrap(next), reader, searchCache, apiPrefix, logger),
+		QueryRangeHandler:         newHandler(queryrange, nil, nil, logger),
 		cacheProvider:             cacheProvider,
 		streamingSearch:           newSearchStreamingGRPCHandler(cfg, o, retryWare.Wrap(next), reader, searchCache, apiPrefix, logger),
 		logger:                    logger,
@@ -238,8 +241,8 @@ func newSearchTagsMiddleware() Middleware {
 	})
 }
 
-// newSpanMetricsMiddleware creates a new frontend middleware to handle search and search tags requests.
-func newSpanMetricsMiddleware() Middleware {
+// newSpanMetricsMiddleware creates a new frontend middleware to handle metrics-generator requests.
+func newMetricsMiddleware() Middleware {
 	return MiddlewareFunc(func(next http.RoundTripper) http.RoundTripper {
 		generatorRT := next
 
@@ -267,4 +270,14 @@ func buildUpstreamRequestURI(originalURI string, params url.Values) string {
 	}
 
 	return uri
+}
+
+func newQueryRangeMiddleware(cfg Config, o overrides.Interface, reader tempodb.Reader, logger log.Logger) Middleware {
+	return MiddlewareFunc(func(next http.RoundTripper) http.RoundTripper {
+		searchRT := NewRoundTripper(next, newQueryRangeSharder(reader, o, cfg.Metrics.Sharder, newSearchProgress, logger))
+
+		return RoundTripperFunc(func(r *http.Request) (*http.Response, error) {
+			return searchRT.RoundTrip(r)
+		})
+	})
 }