+++ modules/frontend/searchsharding.go
@@ -272,10 +272,20 @@ func (s searchSharder) RoundTrip(r *http.Request) (*http.Response, error) {
 		return nil, overallResponse.err
 	}
 
-	if overallResponse.statusCode != http.StatusOK {
-		// translate all non-200s into 500s. if, for instance, we get a 400 back from an internal component
-		// it means that we created a bad request. 400 should not be propagated back to the user b/c
-		// the bad request was due to a bug on our side, so return 500 instead.
+	switch overallResponse.statusCode {
+	case http.StatusOK:
+	case http.StatusTooManyRequests:
+		// Ensure we pass back the 429 status.
+		return &http.Response{
+			StatusCode: overallResponse.statusCode,
+			Header:     http.Header{},
+			Body:       io.NopCloser(strings.NewReader(overallResponse.statusMsg)),
+		}, nil
+	default:
+		// translate all non-200s and non-429s into 500s. if, for instance, we get
+		// a 400 back from an internal component it means that we created a bad
+		// request. 400 should not be propagated back to the user b/c the bad
+		// request was due to a bug on our side, so return 500 instead.
 		return &http.Response{
 			StatusCode: http.StatusInternalServerError,
 			Header:     http.Header{},