+++ modules/generator/processor/localblocks/livetraces.go
@@ -1,15 +1,12 @@
 package localblocks
 
 import (
-	"errors"
 	"hash"
 	"hash/fnv"
 	"time"
 
 	v1 "github.com/grafana/tempo/pkg/tempopb/trace/v1"
 )
-
-var errMaxExceeded = errors.New("asdf")
 
 type liveTrace struct {
 	id        []byte
@@ -25,7 +22,7 @@ type liveTraces struct {
 func newLiveTraces() *liveTraces {
 	return &liveTraces{
 		hash:   fnv.New64(),
-		traces: map[uint64]*liveTrace{},
+		traces: make(map[uint64]*liveTrace),
 	}
 }
 
@@ -39,12 +36,7 @@ func (l *liveTraces) Len() uint64 {
 	return uint64(len(l.traces))
 }
 
-func (l *liveTraces) Push(batch *v1.ResourceSpans, max uint64) error {
-	if len(batch.ScopeSpans) == 0 || len(batch.ScopeSpans[0].Spans) == 0 {
-		return nil
-	}
-
-	traceID := batch.ScopeSpans[0].Spans[0].TraceId
+func (l *liveTraces) Push(traceID []byte, batch *v1.ResourceSpans, max uint64) bool {
 	token := l.token(traceID)
 
 	tr := l.traces[token]
@@ -53,7 +45,7 @@ func (l *liveTraces) Push(batch *v1.ResourceSpans, max uint64) error {
 		// Before adding this check against max
 		// Zero means no limit
 		if max > 0 && uint64(len(l.traces)) >= max {
-			return errMaxExceeded
+			return false
 		}
 
 		tr = &liveTrace{
@@ -64,7 +56,7 @@ func (l *liveTraces) Push(batch *v1.ResourceSpans, max uint64) error {
 
 	tr.Batches = append(tr.Batches, batch)
 	tr.timestamp = time.Now()
-	return nil
+	return true
 }
 
 func (l *liveTraces) CutIdle(idleSince time.Time) []*liveTrace {