+++ modules/overrides/interface.go
@@ -68,6 +68,7 @@ type Interface interface {
 	MaxSearchDuration(userID string) time.Duration
 	DedicatedColumns(userID string) backend.DedicatedColumns
 
-	// API
+	// Management API
 	WriteStatusRuntimeConfig(w io.Writer, r *http.Request) error
+	WriteTenantOverrides(w io.Writer, r *http.Request, tenant string) error
 }